requirements.txt
requests~=2.32.4
dotenv~=0.9.9
python-dotenv~=1.1.1
pip~=22.3.1
wheel~=0.38.4
jinja2~=3.1.6
pytest~=8.4.1
pytest-yaml~=1.2.1
PyYAML~=6.0.2
requests_toolbelt~=1.0.0

jsonschema~=4.24.0
pydantic~=2.11.7
attrs~=25.3.0